added logout and state management
