+++ src/App.vue
@@ -5,7 +5,6 @@
         <a class="navbar-item" href="/#/">
           Bali Guide
         </a>
-
         <a role="button" class="navbar-burger" aria-label="menu" aria-expanded="false">
           <span aria-hidden="true"></span>
           <span aria-hidden="true"></span>
