+++ src/main.js
@@ -2,10 +2,11 @@ import Vue from 'vue';
 import Buefy from 'buefy';
 import 'buefy/dist/buefy.css';
 
-//  import firebase from './firebase';
+import firebase from './firebase';
+import auth from './auth';
 import App from './App.vue';
 import router from './router';
-import store from './vxstore/index';
+import store from './store/index';
 
 Vue.config.productionTip = false;
