+++ src/views/Home.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="home">
     <a class="button is-danger" @click="login()">Login with Google</a>
+    <a class="button is-danger" @click="logout()">Logout with Google</a>
   </div>
 </template>
 
@@ -9,6 +10,6 @@ import { mapActions } from 'vuex';
 
 export default {
   name: 'home',
-  methods: mapActions('auth', ['login']),
+  methods: mapActions('auth', ['login', 'logout']),
 };
 </script>
